requirements.txt
youtrack~=0.1.69
jira~=2.0.0
python-dotenv~=0.15.0
cryptography~=3.4.6
pbr~=5.5.1
setuptools~=54.1.2
six~=1.15.0
requests~=2.25.1
oauthlib~=3.1.0
certifi~=2020.12.5
chardet~=4.0.0
urllib3~=1.26.4
pyparsing~=2.4.7
httplib2~=0.19.0
idna~=2.10
pycparser~=2.20